Algorithm
I added the algorithm to work and it compiles. I change the stocksquery to 100 to deal with the halfchopper/ data loss.

In Update I put place holder numbers and data so that the program would execute.

In stock predict it is currently checking buy and sell and the comment will change to having value that will be obtained by another .cs app